[REFACTOR] 투자자산분류, 매매유형 로직 admin 디렉토리 이동
