+++ src/main/java/com/sysmatic2/finalbe/admin/dto/TradingTypeAdminResponseDto.java
@@ -1,4 +1,4 @@
-package com.sysmatic2.finalbe.strategy.dto;
+package com.sysmatic2.finalbe.admin.dto;
 
 import lombok.Getter;
 import lombok.Setter;
