+++ src/main/java/com/sysmatic2/finalbe/strategy/repository/StrategyRepository.java
@@ -1,7 +1,6 @@
 package com.sysmatic2.finalbe.strategy.repository;
 
 import com.sysmatic2.finalbe.strategy.entity.StrategyEntity;
-import com.sysmatic2.finalbe.strategy.entity.TradingTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
