+++ src/test/java/com/sysmatic2/finalbe/strategy/repository/TradingTypeRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.sysmatic2.finalbe.strategy.repository;
 
-import com.sysmatic2.finalbe.strategy.entity.TradingTypeEntity;
+import com.sysmatic2.finalbe.admin.repository.TradingTypeRepository;
+import com.sysmatic2.finalbe.admin.entity.TradingTypeEntity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
